+++ src/hooks/useNotification.tsx
@@ -1,0 +1,8 @@
+import { useMutation } from '@tanstack/react-query';
+import { createNotification } from '@/api/common/Notification';
+
+export const useNotification = () => {
+  return useMutation({
+    mutationFn: createNotification,
+  });
+};
